Refactored html out of chapter exporter
